refactor: Chat 관련 모델 분리
