+++ RF/Models/Codable/Chat/Sender.swift
@@ -1,0 +1,14 @@
+//
+//  Sender.swift
+//  RF
+//
+//  Created by 이정동 on 2023/09/26.
+//
+
+import Foundation
+
+struct Sender: Codable {
+    var userId: Int?
+    var userName: String?
+    var userImageUrl: String?
+}
