+++ RF/Models/Codable/Enums.swift
@@ -23,7 +23,6 @@ final class EnumFile {
         service.getEnumList()
             .subscribe(
                 onNext: { [weak self] data in
-                    print(data)
                     self?.enumList.accept(data)
                 },onError: { error in
                     print("EnumFile getEnumList error!")
