+++ RF/Models/Realm/RealmChannel.swift
@@ -1,0 +1,13 @@
+//
+//  RealmChannel.swift
+//  RF
+//
+//  Created by 이정동 on 2023/09/26.
+//
+
+import Foundation
+import RealmSwift
+
+class RealmChannel {
+    
+}
