+++ RF/Models/Realm/RealmSchedule.swift
@@ -1,0 +1,13 @@
+//
+//  RealmSchedule.swift
+//  RF
+//
+//  Created by 이정동 on 2023/09/26.
+//
+
+import Foundation
+import RealmSwift
+
+class RealmSchedule {
+    
+}
