+++ RF/Models/Realm/RealmSender.swift
@@ -1,0 +1,13 @@
+//
+//  RealmSender.swift
+//  RF
+//
+//  Created by 이정동 on 2023/09/26.
+//
+
+import Foundation
+import RealmSwift
+
+class RealmSender {
+    
+}
